feat: version API endpoints
